[Frontiers] Restructure the chain of includes.
